feat: registry v2 initial
